requirements.txt
pyobjc-core==9.0.1
pyobjc-framework-AVFoundation==9.0.1
pyobjc-framework-CoreAudio==9.0.1